created ScreenBase wrapped component
