+++ ui/package.json
@@ -5,7 +5,7 @@
   "dependencies": {
     "@alephdata/followthemoney": "2.1.9",
     "@alephdata/react-ftm": "1.10.11",
-    "@blueprintjs/core": "3.31.0",
+    "@blueprintjs/core": "3.33.0",
     "@blueprintjs/datetime": "^3.15.1",
     "@blueprintjs/icons": "3.22.0",
     "@blueprintjs/select": "^3.8.0",
